+++ config/admin/menu.php
@@ -59,7 +59,11 @@ return
             ],
 
 
-
+            [
+                "label"     => "Настройки",
+                "url"       => ["cms/admin-settings"],
+                "img"       => ['\skeeks\cms\modules\admin\assets\AdminAsset', 'images/icons/settings.png'],
+            ],
         ]
     ],