+++ controllers/AdminSettingsController.php
@@ -1,0 +1,107 @@
+<?php
+/**
+ * @author Semenov Alexander <semenov@skeeks.com>
+ * @link http://skeeks.com/
+ * @copyright 2010 SkeekS (СкикС)
+ * @date 27.03.2015
+ */
+namespace skeeks\cms\controllers;
+use skeeks\cms\base\Component;
+use skeeks\cms\modules\admin\controllers\AdminController;
+use yii\helpers\ArrayHelper;
+use yii\web\Response;
+use yii\widgets\ActiveForm;
+
+/**
+ * Class AdminSettingsController
+ * @package skeeks\cms\controllers
+ */
+class AdminSettingsController extends AdminController
+{
+    public function init()
+    {
+        $this->_label                   = "Управление настройками";
+        parent::init();
+    }
+
+    public function actionIndex()
+    {
+        $loadedComponents = [];
+        $loadedForSelect = [];
+        $component = '';
+
+        foreach (\Yii::$app->getComponents(true) as $id => $data)
+        {
+            $loadedComponent = \Yii::$app->get($id);
+            if ($loadedComponent instanceof Component)
+            {
+                $loadedComponents[$loadedComponent->className()]    = $loadedComponent;
+
+                if ($name = ArrayHelper::getValue($loadedComponent->getDescriptorConfig(), 'name'))
+                {
+                    $loadedForSelect[$loadedComponent->className()]     = $name;
+                } else
+                {
+                    $loadedForSelect[$loadedComponent->className()]     = $loadedComponent->className();
+                }
+
+            }
+        }
+
+        if ($componentSelect = \Yii::$app->request->get("component"))
+        {
+            $component = ArrayHelper::getValue($loadedComponents, $componentSelect);
+        }
+
+        if ($component && $component instanceof Component)
+        {
+
+            if (\Yii::$app->request->isAjax && !\Yii::$app->request->isPjax)
+            {
+                $component->load(\Yii::$app->request->post());
+                \Yii::$app->response->format = Response::FORMAT_JSON;
+                return ActiveForm::validate($component);
+            }
+
+            if (\Yii::$app->request->isAjax)
+            {
+                if ($component->load(\Yii::$app->request->post()))
+                {
+                    $this->_setLangAndSite($component);
+                    if ($component->saveDefaultSettings())
+                    {
+                        \Yii::$app->getSession()->setFlash('success', 'Успешно сохранено');
+                    } else
+                    {
+                        \Yii::$app->getSession()->setFlash('error', 'Не удалось сохранить');
+                    }
+
+                } else
+                {
+                    \Yii::$app->getSession()->setFlash('error', 'Не удалось сохранить');
+                }
+
+            }
+        }
+
+        if ($component)
+        {
+            $defaultSettings = $component->fetchDefaultSettings();
+            if ($defaultSettings)
+            {
+                $this->_setLangAndSite($defaultSettings);
+                $values = $defaultSettings->getMultiFieldValue('value');
+                $component->setAttributes($values);
+            }
+        }
+
+
+
+        return $this->render('index', [
+            'loadedComponents'  => $loadedComponents,
+            'loadedForSelect'   => $loadedForSelect,
+            'component'         => $component
+        ]);
+    }
+
+}